requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
chromadb
ollama
psycopg2-binary
pypdf
python-docx
vanna
pytest
httpx
openai